Feature/device logging

+++ SmartHomeManager/SmartHomeManager.Domain/DeviceLoggingDomain/Entities/DeviceLog.cs
@@ -23,10 +23,10 @@ namespace SmartHomeManager.Domain.DeviceLoggingDomain.Entities
         public DateTime DateLogged { get; set; }
 
         [Required]
-        public int DeviceEnergyUsage { get; set; }
+        public double DeviceEnergyUsage { get; set; }
 
         [Required]
-        public int DeviceActivity { get; set; }
+        public double DeviceActivity { get; set; }
 
         [Required]
         public bool DeviceState { get; set; }
@@ -37,4 +37,6 @@ namespace SmartHomeManager.Domain.DeviceLoggingDomain.Entities
         [ForeignKey("DeviceId")]
         public Device Device { get; set; }
     }
+
+
 }
